+++ Sources/F53OSC/F53OSCMessage.h
@@ -3,7 +3,7 @@
 //  F53OSC
 //
 //  Created by Siobhán Dougall on 1/17/11.
-//  Copyright (c) 2011-2025 Figure 53 LLC, https://figure53.com
+//  Copyright (c) 2011-2026 Figure 53 LLC, https://figure53.com
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -51,6 +51,10 @@ NS_ASSUME_NONNULL_BEGIN
 + (BOOL) legalAddressComponent:(nullable NSString *)addressComponent;
 + (BOOL) legalAddress:(nullable NSString *)address;
 + (BOOL) legalMethod:(nullable NSString *)method;
+
+// The address is the first whitespace-delimited component in the string.
+// Optionally returns the remainder (arguments portion) by reference.
++ (nullable NSString *) addressWithString:(NSString *)qscString argumentsString:(out NSString * _Nullable * _Nullable)outArgumentsString;
 
 // Arguments should be space-delimited following the address.
 // Arguments enclosed in quotes are parsed as string arguments.